Added some static casts to the pcp::atom template to satisfy warnings on some compilers.

+++ include/pcp-cpp/atom.hpp
@@ -19,12 +19,12 @@ pmAtomValue atom(const atom_type_type type, ValueType value)
 {
     pmAtomValue atom;
     switch (type) {
-        case PM_TYPE_32:               atom.l   = value; break;
-        case PM_TYPE_U32:              atom.ul  = value; break;
-        case PM_TYPE_64:               atom.ll  = value; break;
-        case PM_TYPE_U64:              atom.ull = value; break;
-        case PM_TYPE_FLOAT:            atom.f   = value; break;
-        case PM_TYPE_DOUBLE:           atom.d   = value; break;
+        case PM_TYPE_32:               atom.l   = static_cast< int32_t>(value); break;
+        case PM_TYPE_U32:              atom.ul  = static_cast<uint32_t>(value); break;
+        case PM_TYPE_64:               atom.ll  = static_cast< int64_t>(value); break;
+        case PM_TYPE_U64:              atom.ull = static_cast<uint32_t>(value); break;
+        case PM_TYPE_FLOAT:            atom.f   = static_cast<float   >(value); break;
+        case PM_TYPE_DOUBLE:           atom.d   = static_cast<double  >(value); break;
       //case PM_TYPE_STRING:           atom.cp  = value; break;
       //case PM_TYPE_AGGREGATE:        atom.vbp = value; break;
       //case PM_TYPE_AGGREGATE_STATIC: atom.vbp = value; break;
